hack(streams): Discard overflowing samples
As discussed in card10/firmware!316, this commit prevents I2C lockup by
discarding overflowing samples instead of blocking until they have been
read.  This is not ideal as the samples read will not be the most recent
ones.  A deeper refactor of the sensor-stream internal API can fix this
in the future.

Cc: @flo_h
Signed-off-by: Rahix <rahix@rahix.de>